messages: Lazy load inline profile pictures (user avatars).
We use the loading attribute to tell the browser to lazy load the
images when they are needed. With this change profile pictures will
only be loaded in they are in the viewport as opposed to previous
behavior where they were loaded when they are added to the DOM and
loaded even though they are not on screen.

We also add a greyish background color to the image tag, which makes
the empty space look nicer in case there is delay in loading the
image or if it fails to load due to internet failure.

Finally, I verfied that there were no issues when sending messages
or reciving message like last time we used lazysizes to do this
(see revert #12112 commit 8a15b9ee87d42ba8b4f810c5aa4e6f9ae45b6b25).
The two issues were profile pictures flashes (or flickering) when
sending messages or when reciving new messages when they are
dynamically rendered.